fix two identical player bug and flash block

+++ js/init.coffee
@@ -79,8 +79,28 @@ window.collaborator_left_callback = (evt)->
 
 
 Nimbus.Auth.set_app_ready(()->
-	sync_players_on_callback()
+	search = localStorage['doc_id']
+	if search and search isnt c_file.id
+		load_new_file(search,()->
+			console.log 'loading new file'
+			sync_players_on_callback()
+		,(e)->
+			if e.type is gapi.drive.realtime.ErrorType.TOKEN_REFRESH_REQUIRED
+				authorizer.authorize()
+			else if e.type is gapi.drive.realtime.ErrorType.CLIENT_ERROR
+				if localStorage['doc_id']
+					localStorage.clear()
+					location.reload()
+				else
+					alert "An Error happened: " + e.message
+			else
+				console.log 'Unknown error:'+e.message
+		)
+		return
+	else
+		sync_players_on_callback()
 	return
+
 )
 
 # sync_on_start
@@ -89,24 +109,14 @@ window.sync_players_on_callback = ()->
 	if Nimbus.Auth.authorized()
 		url = location.pathname+'?'+c_file.id
 		window.history.pushState("Game Started", "Nimbus Tetris", url)
+		localStorage['doc_id'] = c_file.id
 		collabrators = doc.getCollaborators()
 		for one in collabrators
 			if one.isMe
 				localStorage['current'] = JSON.stringify(one)
 				break
-		if window.new_game is true
-			window.new_game = false
-			return process_game_data()
-		
-		if app_files and app_files.length < 2
-			$('.mask').hide()
-			process_game_data()
-		else
-			# parse user and game on panel
-			list_games()
-			$('#login').hide();
-			$('.mask .panel').slideDown()
-
+		process_game_data()
+		$('.mask').fadeOut()
 window.list_games = ()->
 	html = ''
 	profile = ''
@@ -270,19 +280,24 @@ window.remove_file = (file_id,evt)->
 $ ()->
 	$('.panel .list').on('click',(evt)->
 		file_id = $(evt.target).data('id')
-		if $(evt.target)[0].tagName is 'P'
+		if $(evt.target)[0].tagName is 'P' and c_file.id isnt file_id
 			# remove file
 			remove_file(file_id,evt)
-		else if $(evt.target)[0].tagName is 'A'
+		else if $(evt.target)[0].tagName is 'A' and c_file.id isnt file_id
 			erase_indexedDB(()->
 				window.controllers.new_game() if window.controllers
+				try
+					doc.close()
+				catch e
+					console.log e
 				Nimbus.Client.GDrive.switch_to_app_file_real(file_id,()->
 					url = location.pathname+'?'+c_file.id
 					window.history.pushState("New Game Loaded", "Nimbus Tetris", url)
-					process_game_data()
+					localStorage['doc_id'] = c_file.id
 				)
 				
 				$('.mask').fadeOut()
+				return
 			)
 		else
 			$('.mask').fadeOut() if window.controllers
@@ -306,6 +321,11 @@ $ ()->
 				if Nimbus.dictModel?
 					for k, v of Nimbus.dictModel
 						v.records = {}
+				try
+					doc.close()
+				catch e
+					console.log e
+				
 				gapi.drive.realtime.load(data.id, onFileLoaded, initializeModel);
 			)
 			return
